# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/households": {
      "get": {
        "operationId": "HouseholdsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Households de l'utilisateur, avec son rôle dans chacun."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste les households dont l'utilisateur courant est membre.",
        "tags": [
          "households"
        ]
      },
      "post": {
        "operationId": "HouseholdsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Household créé."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée un household ; le créateur en devient OWNER.",
        "tags": [
          "households"
        ]
      }
    },
    "/households/{householdId}": {
      "get": {
        "operationId": "HouseholdsController_getById",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Household trouvé."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retourne les informations d'un household.",
        "tags": [
          "households"
        ]
      },
      "patch": {
        "operationId": "HouseholdsController_rename",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Household renommé."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Renomme un household (réservé à OWNER/ADMIN).",
        "tags": [
          "households"
        ]
      },
      "delete": {
        "operationId": "HouseholdsController_remove",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Household supprimé."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprime un household (réservé à OWNER).",
        "tags": [
          "households"
        ]
      }
    },
    "/households/{householdId}/members": {
      "get": {
        "operationId": "HouseholdsController_listMembers",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membres du household, avec leur rôle."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste les membres d'un household.",
        "tags": [
          "households"
        ]
      }
    },
    "/households/{householdId}/members/{userId}": {
      "patch": {
        "operationId": "HouseholdsController_updateMemberRole",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rôle mis à jour."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change le rôle d'un membre (réservé à OWNER/ADMIN).",
        "tags": [
          "households"
        ]
      },
      "delete": {
        "operationId": "HouseholdsController_removeMember",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Membre retiré."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retire un membre du household (réservé à OWNER/ADMIN).",
        "tags": [
          "households"
        ]
      }
    },
    "/households/{householdId}/leave": {
      "post": {
        "operationId": "HouseholdsController_leave",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Household quitté."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fait quitter le household courant à l’utilisateur authentifié.",
        "tags": [
          "households"
        ]
      }
    },
...
    "/invitations/{token}/accept": {
      "post": {
        "operationId": "InvitationsPublicController_accept",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Invitation acceptée, utilisateur ajouté au household."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accepte une invitation par son jeton (l'email du compte connecté doit correspondre à l'invitation).",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/{invitationId}/revoke": {
      "post": {
        "operationId": "InvitationsPublicController_revoke",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Invitation révoquée."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Révoque une invitation en attente (réservé à OWNER/ADMIN du household concerné).",
        "tags": [
          "invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Notre Nid API",
    "description": "API de gestion de collection partagée pour Notre Nid.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateHouseholdDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Notre nid"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateHouseholdDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Notre nid"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}